+++ packages/iov-bns/src/bnsconnection.ts
@@ -398,7 +398,7 @@ export class BnsConnection implements BcpAtomicSwapConnection {
   /**
    * Emits currentState (getSwap) as a stream, then sends updates for any matching swap
    *
-   * This includes an open swap beind claimed/expired as well as a new matching swap being offered
+   * This includes an open swap beind claimed/aborted as well as a new matching swap being offered
    */
   public watchSwaps(query: AtomicSwapQuery): Stream<AtomicSwap> {
     // we need to combine them all to see all transactions that affect the query